% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{add_base_info}
\alias{add_base_info}
\title{Function to add the base information to mutation_status}
\usage{
add_base_info(reads_df)
}
\description{
Function to add the base information to mutation_status
}
