% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{adjust_quality_for_deletions}
\alias{adjust_quality_for_deletions}
\title{Function to add '!' for bases that were deleted in the qual columns}
\usage{
adjust_quality_for_deletions(md_tag, clip_qual)
}
\description{
Function to add '!' for bases that were deleted in the qual columns
}
