% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{adjust_ref_fragments}
\alias{adjust_ref_fragments}
\title{Helper function for downsampling reference base fragments in motifs}
\usage{
adjust_ref_fragments(gr, ref_type, downsample_ref = FALSE)
}
\description{
Helper function for downsampling reference base fragments in motifs
}
