% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{adjust_sequence_for_deletions}
\alias{adjust_sequence_for_deletions}
\title{Function to add 'D' for bases that were deleted in the seq columns}
\usage{
adjust_sequence_for_deletions(md_tag, clip_seq)
}
\description{
Function to add 'D' for bases that were deleted in the seq columns
}
