% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{annotate_cnv_mut}
\alias{annotate_cnv_mut}
\title{Helper function for plotting CNV and mutational data}
\usage{
annotate_cnv_mut(
  p,
  olap_df,
  segment_line_size,
  output_file,
  ggsave_params,
  ...
)
}
\description{
Helper function for plotting CNV and mutational data
}
