% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{apply_to_paired_reads}
\alias{apply_to_paired_reads}
\title{Wrapper function applying mm_del_pos to each row with _first/_last suffix}
\usage{
apply_to_paired_reads(data_frame)
}
\description{
Wrapper function applying mm_del_pos to each row with _first/_last suffix
}
