% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{apply_to_paired_reads_ins}
\alias{apply_to_paired_reads_ins}
\title{Function to extract insertion information from paired reads}
\usage{
apply_to_paired_reads_ins(data_frame)
}
\description{
Function to extract insertion information from paired reads
}
