% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBam.R
\name{bam_to_galp2}
\alias{bam_to_galp2}
\title{Function to read bam file}
\usage{
bam_to_galp2(
  bamfile,
  use_names = TRUE,
  param = galp_param,
  chromosome_to_keep = FALSE,
  strand_mode = 1,
  genome = NA_character_,
  ...
)
}
\description{
Function to read bam file
}
