% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{bam_to_galp_mut}
\alias{bam_to_galp_mut}
\title{Function to subset alignments to mutational positions}
\usage{
bam_to_galp_mut(
  bamfile,
  use_names = TRUE,
  param = galp_param,
  chromosome_to_keep = FALSE,
  strand_mode = 1,
  genome = NA_character_
)
}
\description{
Function to subset alignments to mutational positions
}
