% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCNV.R
\name{callCNV}
\alias{callCNV}
\title{call copy number variation using QDNAseq utilities}
\usage{
callCNV(bamfile, bin_size = 1000, genome_label = "hg19")
}
\arguments{
\item{bamfile}{}

\item{bin_size}{}

\item{genome_label}{}
}
\value{
QDNAseqCopyNumbers object
}
\description{
call copy number variation using QDNAseq utilities
}
