% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMotif.R
\name{callMotif}
\alias{callMotif}
\title{Call Motif}
\usage{
callMotif(
  fragment_obj,
  genome_label = "hg19",
  motif_type = "s",
  motif_length = 3L,
  ...
)
}
\arguments{
\item{fragment_obj}{}

\item{genome_label}{}

\item{motif_type}{}

\item{motif_length}{}

\item{...}{}
}
\value{
tibble object
}
\description{
Call Motif
}
