% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callTrinucleotide.R
\name{callTrinucleotide}
\alias{callTrinucleotide}
\title{Call trinucleotides and summarise the cfDNA information
for each target mutation locus.}
\usage{
callTrinucleotide(frag_obj_mut)
}
\arguments{
\item{frag_obj_mut}{GRanges object containing fragment and mutation data.}
}
\value{
dataframe with summarised mutational and trinucleotide data.
}
\description{
This function processes a GRanges object, summarizing cfDNA fragment
information for each target mutation locus.
}
\details{
It annotates each mutation locus with the number
and type of supporting fragments according
to their read-pair overlap status.

The median fragment length is also annotated
for each read-pair overlap type.

Locus-based consensus mutation is determined
by selecting the most frequent mismatch type,
with priority given to concordant read-pair mutations (CO_MUT)
followed by single-read mutations (SO_MUT).
The consensus mutations are then used to derive
the trinucleotide substitution types (SBS96).
}
\examples{
trinuc_df <-  callTrinucleotide(frag_obj_mut)
}
