% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{check_mutation_in_metadata}
\alias{check_mutation_in_metadata}
\title{Function to check for mutational metadata in the GRanges object}
\usage{
check_mutation_in_metadata(frag_obj)
}
\description{
Function to check for mutational metadata in the GRanges object
}
