% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{check_mutation_status}
\alias{check_mutation_status}
\title{Function to check for mutant base fragments in a dataframe}
\usage{
check_mutation_status(reads_df)
}
\description{
Function to check for mutant base fragments in a dataframe
}
