% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{check_mutfile_columns}
\alias{check_mutfile_columns}
\title{Function to check Mutation file columns}
\usage{
check_mutfile_columns(df)
}
\arguments{
\item{df}{A dataframe representing the mutation file.}
}
\description{
Function to check Mutation file columns
}
