% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{clip_read_pair}
\alias{clip_read_pair}
\title{Function to remove soft clipped bases to match the MD tags
#' @param reads_df A dataframe containing read pairs with 'cigar.first' and
'cigar.last' columns that may include soft clipping indicators.}
\usage{
clip_read_pair(reads_df = NULL)
}
\description{
Function to remove soft clipped bases to match the MD tags
#' @param reads_df A dataframe containing read pairs with 'cigar.first' and
'cigar.last' columns that may include soft clipping indicators.
}
