% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{create_empty_galp}
\alias{create_empty_galp}
\title{Create an Empty GAlignmentPairs Object}
\usage{
create_empty_galp()
}
\value{
An empty GAlignmentPairs object.
}
\description{
Initializes an empty GAlignmentPairs object with two GAlignments objects,
one for each pair member, and combines them.
}
