% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePath.R
\name{examplePath}
\alias{examplePath}
\title{Get path to cfDNAPro example folder.}
\usage{
examplePath(data = NULL)
}
\arguments{
\item{data}{Name of data set. Such as "groups" or "step6".
If `NULL`, the path of extdata folder will be returned.}
}
\value{
A string. (i.e. the path.)
}
\description{
cfDNAPro package has sample files in `inst/extdata`
directory. This function helps get the path to the data.
}
\examples{
examplePath()
examplePath("groups_picard")
examplePath("step6")
}
