% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{get_genome_reference}
\alias{get_genome_reference}
\title{Function to determine the correct BSgenome based on the seqinfo result}
\usage{
get_genome_reference(frag_obj_mut)
}
\description{
Function to determine the correct BSgenome based on the seqinfo result
}
