% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{get_highest_column}
\alias{get_highest_column}
\title{Helper function which selects consensus mutation type based on counts}
\usage{
get_highest_column(row)
}
\description{
Helper function which selects consensus mutation type based on counts
}
