% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{get_trinucleotide}
\alias{get_trinucleotide}
\title{Helper function which generates SBS 96 trinucleotide channel information}
\usage{
get_trinucleotide(merged_table_df, genome)
}
\description{
Helper function which generates SBS 96 trinucleotide channel information
}
