% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{insertion_pos}
\alias{insertion_pos}
\title{Function to parse CIGAR strings and extract insertions}
\usage{
insertion_pos(
  unique_read_pair_id,
  chr,
  start_pos,
  end_pos,
  read_strand,
  nm_tag,
  md_tag,
  cigar_string,
  clip_seq,
  clip_qual,
  mapq,
  which_locus
)
}
\description{
Function to parse CIGAR strings and extract insertions
}
