% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{integrate_motif_mut}
\alias{integrate_motif_mut}
\title{Helper function to process motif and mutation data}
\usage{
integrate_motif_mut(
  frag_obj,
  downsample_ref,
  ref_type,
  bsgenome_obj,
  motif_type,
  motif_length
)
}
\description{
Helper function to process motif and mutation data
}
