% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{mm_del_pos}
\alias{mm_del_pos}
\title{Function to add mismatch and deletion info as separate columns for readpairs}
\usage{
mm_del_pos(
  unique_read_pair_id,
  chr,
  start_pos,
  end_pos,
  read_strand,
  nm_tag,
  md_tag,
  cigar_string,
  clip_seq,
  clip_qual,
  mapq,
  which_locus
)
}
\description{
Function to add mismatch and deletion info as separate columns for readpairs
}
