% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{plot_length_mut}
\alias{plot_length_mut}
\title{Function which generates fragment length plot with mutational data}
\usage{
plot_length_mut(x, ylim, output_file, ggsave_params)
}
\description{
Function which generates fragment length plot with mutational data
}
