% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{plot_motif_mut}
\alias{plot_motif_mut}
\title{Helper function for plotting fragment motif and mutational data
Helper function for plotting fragment motif and mutational data}
\usage{
plot_motif_mut(
  x,
  ylim,
  x_title,
  plot_type,
  bar_color,
  motif_levels,
  output_file,
  ggsave_params
)
}
\description{
Helper function for plotting fragment motif and mutational data
Helper function for plotting fragment motif and mutational data
}
