% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{plotCNV}
\alias{plotCNV}
\title{plot copy number profile}
\usage{
plotCNV(
  x,
  gene_to_highlight = list(ENTREZID = NULL, ENSEMBL = NULL, SYMBOL = c("EGFR", "ERBB2")),
  genome = "hg19",
  ylim = c(-2, 2),
  chromosome = c(seq(1, 22, 1), "X"),
  point_color = c(Loss = "royalblue", Deletion = "darkblue", Neutral = "darkgrey",
    Amplification = "darkorange", Gain = "orange3"),
  x_title = "Chromosome",
  y_title = "Log2 Ratio",
  point_size = 0.3,
  point_alpha = 0.9,
  chr_edge_color = "black",
  chr_edge_line_size = 0.2,
  chr_edge_alpha = 0.8,
  chr_edge_type = "dotted",
  segment_color = "red",
  segment_alpha = 1,
  segment_line_end = "round",
  segment_line_size = 0.75,
  legend_position = "none",
  x_axis_expand = c(0.1, 0.1),
  y_axis_expand = c(0, 0),
  ...
)
}
\arguments{
\item{x}{QDNAseqCopyNumbers object}

\item{gene_to_highlight}{A named list.}

\item{genome}{string. hg19 or hg38}

\item{ylim}{vector. default is c(-2, 2).}

\item{chromosome}{vector, default is c(seq(1, 22, 1), "X").}

\item{point_color}{named vector}

\item{x_title}{string.}

\item{y_title}{string.}

\item{point_size}{numerical. default is 0.3.}

\item{point_alpha}{numerical. default is 0.9.}

\item{chr_edge_color}{string. default is "black".}

\item{chr_edge_line_size}{numerical. default is 0.2.}

\item{chr_edge_alpha}{numerical. default is 0.8.}

\item{chr_edge_type}{string. default is "dotted".}

\item{segment_color}{string. default is "red".}

\item{segment_alpha}{numerical. default is 1.}

\item{segment_line_end}{string. default is "round".}

\item{segment_line_size}{numerical. default is 0.75.}

\item{legend_position}{string. default is "none", which mean no legends.}

\item{x_axis_expand}{numerical vector. default is c(0.1, 0.1).}

\item{y_axis_expand}{numerical vector. default is c(0, 0).}

\item{...}{Other params to geom_txt_repel()}
}
\value{
This function returns ggplot2 object.
}
\description{
plot copy number profile
}
\examples{
\dontrun{

p <- plotCNV(bamfile = "/path/to/bamfile.bam")
}

}
\author{
Haichao Wang
}
