% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{plotTrinucData}
\alias{plotTrinucData}
\title{Function which runs the the plotting operations for SBS96 profile}
\usage{
plotTrinucData(
  count_df,
  ylim = ylim,
  show_overlap_type = TRUE,
  plot_title = "Trinucleotide Profile",
  y_axis_title = "Proportion of Single Base Substitutions",
  draw_x_axis_labels = TRUE,
  draw_y_axis_labels = TRUE,
  draw_y_axis_title = TRUE,
  output_file = output_file,
  ggsave_params = ggsave_params
)
}
\description{
Function which runs the the plotting operations for SBS96 profile
}
