% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_bam_file}
\alias{process_bam_file}
\title{Function to process BAM files based on mutational or general alignment data}
\usage{
process_bam_file(
  bamfile,
  mutation_file = NULL,
  mut_fragments_only = FALSE,
  use_names = FALSE,
  chromosome_to_keep = NULL,
  strand_mode = 1,
  genome_name,
  galp_flag,
  galp_what,
  galp_tag,
  galp_mapqFilter
)
}
\description{
Function to process BAM files based on mutational or general alignment data
}
