% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_cnv_mut}
\alias{process_cnv_mut}
\title{Helper function to process mutational information for CNV analysis}
\usage{
process_cnv_mut(olap, frag_obj_mut)
}
\description{
Helper function to process mutational information for CNV analysis
}
