% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_columns}
\alias{process_columns}
\title{Internal Functions for mutational processing}
\usage{
process_columns(df, suffix)
}
\description{
Internal Functions for mutational processing
}
