% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_duplicate_read_pairs}
\alias{process_duplicate_read_pairs}
\title{Function to process duplicate read-pair IDs in galp/granges objects
and generate a dataframe for later processing}
\usage{
process_duplicate_read_pairs(genomic_data, mut_fragments_only)
}
\description{
Function to process duplicate read-pair IDs in galp/granges objects
and generate a dataframe for later processing
}
