% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_insertions}
\alias{process_insertions}
\title{Function to process insertion information}
\usage{
process_insertions(reads_df)
}
\description{
Function to process insertion information
}
