% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_length_mut}
\alias{process_length_mut}
\title{Helper function to process mutation-related information}
\usage{
process_length_mut(frag_obj, ref_type, downsample_ref)
}
\description{
Helper function to process mutation-related information
}
