% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_mutation_fragments}
\alias{process_mutation_fragments}
\title{Function to process and curate mutational fragment-level information}
\usage{
process_mutation_fragments(
  bamfile = bamfile,
  genome = genome,
  galp = galp,
  mutation_file = mutation_file,
  frag = frag,
  galp_bqFilter = galp_bqFilter,
  chromosome_to_keep = chromosome_to_keep,
  mut_fragments_only = mut_fragments_only
)
}
\description{
Function to process and curate mutational fragment-level information
}
