% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{process_mutation_status}
\alias{process_mutation_status}
\title{Function to derive information about the mutation locus from each read-pair}
\usage{
process_mutation_status(reads_df)
}
\description{
Function to derive information about the mutation locus from each read-pair
}
