% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMotif.R
\name{processMotif}
\alias{processMotif}
\title{Helper function to process the motif data}
\usage{
processMotif(frag_obj, bsgenome_obj, motif_type, motif_length)
}
\description{
Helper function to process the motif data
}
