% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{processTrinucleotideData}
\alias{processTrinucleotideData}
\title{Function which generates complete SBS96 profiles with readpair overlap info
Function which generates complete SBS96 profiles with readpair overlap info}
\usage{
processTrinucleotideData(
  trinuc_df,
  exclude_if_type_present = NULL,
  retain_if_type_present = NULL,
  remove_type = NULL,
  normalize_counts = TRUE
)
}
\description{
Function which generates complete SBS96 profiles with readpair overlap info
Function which generates complete SBS96 profiles with readpair overlap info
}
