% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bam_insert_metrics.R
\name{read_bam_insert_metrics}
\alias{read_bam_insert_metrics}
\title{Calculate insert sizes from a curated GRanges object}
\usage{
read_bam_insert_metrics(
  bamfile,
  chromosome_to_keep = paste0("chr", 1:22),
  strand_mode = 1,
  genome_label = "hg19",
  outdir = NA,
  isize_min = 1L,
  isize_max = 1000L,
  ...
)
}
\arguments{
\item{bamfile}{The bam file name.}

\item{chromosome_to_keep}{Should be a character vector containing the 
seqnames to be kept in the GRanges object. 
Default is paste0("chr", 1:22).}

\item{strand_mode}{Usually the strand_mode  = 1 means the First read is 
aligned to positive strand. Details please see GenomicAlignments docs.}

\item{genome_label}{The Genome you used in the alignment. 
Should be "hg19" or "hg38". Default is "hg19".}

\item{outdir}{The path for saving rds file. Default is NA, i.e. not saving.}

\item{isize_min}{min fragment lenght to keep, default is 1L.}

\item{isize_max}{max fragment length to keep, default is 1000L.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Calculate insert sizes from a curated GRanges object
}
