% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{read_mutation_file}
\alias{read_mutation_file}
\title{Function to read in the .tsv mutation file}
\usage{
read_mutation_file(mutation_file)
}
\arguments{
\item{mutation_file}{A string path to the TSV file containing mutation data.
The file should contain columns named 'chr', 'pos', 'ref', and 'alt'.}
}
\description{
Function to read in the .tsv mutation file
}
