% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{remove_insertion_bases}
\alias{remove_insertion_bases}
\title{Function to remove insertion bases from the seq columns}
\usage{
remove_insertion_bases(cigar, clip_seq)
}
\description{
Function to remove insertion bases from the seq columns
}
