% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{remove_insertion_qual_bases}
\alias{remove_insertion_qual_bases}
\title{Function to remove insertion quality scores from the base quality columns}
\usage{
remove_insertion_qual_bases(cigar, clip_qual)
}
\description{
Function to remove insertion quality scores from the base quality columns
}
