% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{select_and_normalize}
\alias{select_and_normalize}
\title{Helper function to select and optionally normalize REF fragment counts}
\usage{
select_and_normalize(frag_obj, pattern, downsample_ref = FALSE, target_count)
}
\description{
Helper function to select and optionally normalize REF fragment counts
}
