% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{stratify_mutation_status}
\alias{stratify_mutation_status}
\title{Process mutation status and split into components}
\usage{
stratify_mutation_status(gr_df)
}
\description{
Process mutation status and split into components
}
