% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{summarize_mutational_data}
\alias{summarize_mutational_data}
\title{Helper function to summarize mutational data}
\usage{
summarize_mutational_data(gr_df)
}
\description{
Helper function to summarize mutational data
}
