% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBam.R
\name{summarizeBam}
\alias{summarizeBam}
\title{Summarise descriptive Bam stats}
\usage{
summarizeBam(
  bamfile = NULL,
  total_count = TRUE,
  total_mapped_count = TRUE,
  chrM_count = TRUE,
  duplicate_count = TRUE,
  coverage_by_mapped = TRUE,
  genome_length_bp = 3.2e+09,
  gc_metrics = FALSE,
  loci_coverage_metrics = FALSE,
  customized_count = FALSE,
  customized_count_mapqFilter = NA_integer_,
  customized_count_isPaired = NA,
  customized_count_isProperPair = NA,
  customized_count_isUnmappedQuery = NA,
  customized_count_hasUnmappedMate = NA,
  customized_count_isMinusStrand = NA,
  customized_count_isMateMinusStrand = NA,
  customized_count_isFirstMateRead = NA,
  customized_count_isSecondMateRead = NA,
  customized_count_isSecondaryAlignment = NA,
  customized_count_isNotPassingQualityControls = NA,
  customized_count_isDuplicate = NA,
  customized_count_isSupplementaryAlignment = NA,
  ...
)
}
\arguments{
\item{bamfile}{Bam file}

\item{total_count}{Boolean. default = TRUE, which means calculating the total 
number of reads.}

\item{total_mapped_count}{Boolean. Default = TRUE, which mean calculating the
number of mapped reads, and these reads must have mate reads.}

\item{chrM_count}{}

\item{duplicate_count}{}

\item{coverage_by_mapped}{}

\item{genome_length_bp}{}

\item{customized_count}{}

\item{customized_count_mapqFilter}{}

\item{customized_count_isPaired}{}

\item{customized_count_isProperPair}{}

\item{customized_count_isUnmappedQuery}{}

\item{customized_count_hasUnmappedMate}{}

\item{customized_count_isMinusStrand}{}

\item{customized_count_isMateMinusStrand}{}

\item{customized_count_isFirstMateRead}{}

\item{customized_count_isSecondMateRead}{}

\item{customized_count_isSecondaryAlignment}{}

\item{customized_count_isNotPassingQualityControls}{}

\item{customized_count_isDuplicate}{}

\item{customized_count_isSupplementaryAlignment}{}

\item{...}{}
}
\value{
tibble object (i.e. a dataframe)
}
\description{
Summarise descriptive Bam stats
}
\examples{
\dontrun{

 summarizeBam(bamfile = "/path/to/bamfile.bam")
}

}
\author{
Haichao Wang
}
