% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{update_consensus_mismatch}
\alias{update_consensus_mismatch}
\title{Function to update consensus_mismatch based on the highest priority match}
\usage{
update_consensus_mismatch(merged_table_df, gr_df, mapping)
}
\description{
Function to update consensus_mismatch based on the highest priority match
}
