% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalMutFunctions.R
\name{update_mdi_columns}
\alias{update_mdi_columns}
\title{Function to obtain updated mismatch and deletion information}
\usage{
update_mdi_columns(reads_df)
}
\description{
Function to obtain updated mismatch and deletion information
}
