% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callCNV.R
\name{callCNV}
\alias{callCNV}
\title{Call Copy Number Variation Using QDNAseq Utilities}
\usage{
callCNV(bamfile, bin_size = 1000, genome_label = "hg19")
}
\arguments{
\item{bamfile}{A string specifying the path to a single BAM file.}

\item{bin_size}{A numerical value indicating the size of the bins (default is 1000).}

\item{genome_label}{A string specifying the genome label (default is "hg19").}
}
\value{
A `QDNAseqCopyNumbers` object containing the copy number variation results.
}
\description{
This function leverages QDNAseq to process BAM files for CNV analysis.
It applies a sequence of operations including binning, filtering,
correcting, normalizing, smoothing, segmenting, and calling to
generate a QDNAseqCopyNumbers object.
}
\examples{
callCNV(bamfile = "path/to/your/file.bam", bin_size = 1000, genome_label = "hg19")

}
