% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callLength.R
\name{callLength}
\alias{callLength}
\title{call fragment length}
\usage{
callLength(fragment_obj, isize_min = 1L, isize_max = 1000L, ...)
}
\arguments{
\item{fragment_obj}{}

\item{isize_min}{}

\item{isize_max}{}

\item{...}{}
}
\value{
a tibble
}
\description{
call fragment length
}
