% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampleBam.R
\name{downsampleBam}
\alias{downsampleBam}
\title{Randomly Downsample BAM File to a Target Depth}
\usage{
downsampleBam(
  bamfile,
  genome = NA_character_,
  input_depth = NULL,
  input_mapped_reads_count = NULL,
  output_depth = 0.1,
  output_type = c("bam", "rds"),
  output_dir = NULL,
  return_result = TRUE,
  ...
)
}
\arguments{
\item{bamfile}{Path to a BAM file.}

\item{genome}{Supported values: "hg19", "hg38", "GRCh38", or NA. Sets seqinfo.}

\item{input_depth}{Target depth, inferred if not supplied.}

\item{input_mapped_reads_count}{Mapped reads count to infer depth if not supplied.}

\item{output_depth}{Target output depth after downsampling, default is 0.1.}

\item{output_type}{Output file formats, default is c("bam", "rds").}

\item{output_dir}{Directory for output files, defaults to input file directory.}

\item{return_result}{If TRUE (default), returns result as an R object.}

\item{...}{Additional parameters for `bam_to_galp2` function.}
}
\value{
`GAlignmentPairs` object with the downsampled reads if return_result.
}
\description{
Downsampling a BAM file to achieve a specified sequencing depth. Exports the
data as a BAM file and/or an RDS file containing a `GAlignmentPairs` object.
}
\examples{
\dontrun{
  # Downsample a BAM file to 10\% of its original depth
  result <- downsampleBam("/path/to/your.bam", genome = "hg38",
                          output_depth = 0.1, output_type = c("bam", "rds"),
                          return_result = TRUE)
}
}
