% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{plotCNV}
\alias{plotCNV}
\title{plot copy number profile}
\usage{
plotCNV(
  x,
  ylim = c(-2, 2),
  chromosome = c(seq(1, 22, 1), "X"),
  point_color = c(Loss = "royalblue", Deletion = "darkblue", Neutral = "darkgrey",
    Amplification = "darkorange", Gain = "orange3"),
  x_title = "Chromosome",
  y_title = "Log2 Ratio",
  point_size = 0.3,
  point_alpha = 0.9,
  chr_edge_color = "black",
  chr_edge_line_size = 0.2,
  chr_edge_alpha = 0.8,
  chr_edge_type = "dotted",
  segment_color = "red",
  segment_alpha = 1,
  segment_line_end = "round",
  segment_line_size = 0.75,
  legend_position = "none",
  x_axis_expand = c(0.1, 0.1),
  y_axis_expand = c(0, 0),
  ...
)
}
\arguments{
\item{x}{}

\item{ylim}{}

\item{chromosome}{}

\item{point_color}{}

\item{x_title}{}

\item{y_title}{}

\item{point_size}{}

\item{point_alpha}{}

\item{chr_edge_color}{}

\item{chr_edge_line_size}{}

\item{chr_edge_alpha}{}

\item{chr_edge_type}{}

\item{segment_color}{}

\item{segment_alpha}{}

\item{segment_line_end}{}

\item{segment_line_size}{}

\item{legend_position}{}

\item{x_axis_expand}{}

\item{y_axis_expand}{}

\item{...}{}
}
\value{
This function returns ggplot2 object.
}
\description{
plot copy number profile
}
