% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLength.R
\name{plotLength}
\alias{plotLength}
\title{plot fragment length profile}
\usage{
plotLength(
  x,
  plot_type = c("Fraction"),
  xlim = c(30, 500),
  ylim,
  x_breaks,
  x_labels,
  vline,
  add_vline = TRUE,
  line_color = "grey1",
  line_size = 0.6,
  line_alpha = 0.8,
  vline_color = "grey",
  vline_type = "dashed",
  vline_size = 0.7,
  area_highlight = NULL,
  ...
)
}
\arguments{
\item{x}{}

\item{plot_type}{}

\item{xlim}{}

\item{ylim}{}

\item{x_breaks}{}

\item{x_labels}{}

\item{vline}{}

\item{add_vline}{}

\item{line_color}{}

\item{line_size}{}

\item{line_alpha}{}

\item{vline_color}{}

\item{vline_type}{}

\item{vline_size}{}

\item{area_highlight}{}

\item{...}{}
}
\value{
ggplot2 object
}
\description{
plot fragment length profile
}
