% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotif.R
\name{plotMotif}
\alias{plotMotif}
\title{plot motif count or fraction as bar plot}
\usage{
plotMotif(
  x,
  ylim,
  x_title = "Motif",
  plot_type = c("Fraction"),
  bar_color = c(A = "cornflowerblue", C = "darksalmon", G = "palevioletred2", T =
    "forestgreen"),
  motif_levels = c("C", "G", "A", "T"),
  ...
)
}
\arguments{
\item{x}{}

\item{ylim}{}

\item{x_title}{}

\item{plot_type}{}

\item{bar_color}{}

\item{motif_levels}{}

\item{...}{}
}
\value{
ggplot2 object
}
\description{
plot motif count or fraction as bar plot
}
